Adding a cat picture
A picture of a cat.  I found it on google.  I think I will try committing this directly to master.  The plan is to then go back to the branch I was working on then insert this picture into the main page.  At least until I figure out what I actually want to do with this blog.